Position is now doubles The Location Service now allows API Key validation

+++ GWvW_Overlay-Location-Server-Contracts/DataContracts.cs
@@ -33,7 +33,7 @@ namespace GWvW_Overlay_Location_Server_Contracts
     public class Position
     {
         [DataMember(IsRequired = true)]
-        public int X, Y;
+        public double X, Y;
 
         [DataMember(IsRequired = true)]
         public int MapId;
